requirements.txt
asgiref==3.6.0
attrs==19.3.0
Automat==0.8.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
#dbus-python==1.2.16
distro==1.4.0
Django==4.1.7
entrypoints==0.3
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
# onboard==1.4.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.9.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.10.0
#pycups==1.9.73
#PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
# python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
# python-debian===0.1.36ubuntu1
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
# sos==4.2
sqlparse==0.4.3
ssh-import-id==5.10
# systemd-python==234
tomli==2.0.1
Twisted==18.9.0
# ubuntu-advantage-tools==27.6
# ubuntu-drivers-common==0.0.0
# ufw==0.36
# unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
# xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
django-cors-headers
djangorestframework
npm
